Move multiline comment to SQR.functions
